@@ -1,20 +1,26 @@
 <?php
 
-namespace Locastic\ActivityLog\Message;
+namespace Locastic\Loggastic\Message;
 
-use Locastic\ActivityLog\Enum\ActivityLogAction;
+use Locastic\Loggastic\Enum\ActivityLogAction;
+use Locastic\Loggastic\Util\ClassUtils;
 
-class UpdateActivityLogMessage
+class UpdateActivityLogMessage implements UpdateActivityLogMessageInterface
 {
     private \DateTime $dateTime;
-    private $updatedItem;
+    private object $updatedItem;
     private string $actionName;
+    private bool $createLogWithoutChanges;
+    private ?array $userInfo = null;
+    private ?string $requestUrl = null;
+    private array $normalizedItem = [];
 
-    public function __construct($updatedItem, ?string $actionName = null)
+    public function __construct(object $updatedItem, ?string $actionName = null, bool $createLogWithoutChanges = false)
     {
         $this->dateTime = new \DateTime();
         $this->updatedItem = $updatedItem;
-        $this->actionName = $actionName ?? ActivityLogAction::$EDITED;
+        $this->actionName = $actionName ?? ActivityLogAction::EDITED;
+        $this->createLogWithoutChanges = $createLogWithoutChanges;
     }
 
     public function getDateTime(): \DateTime
@@ -27,7 +33,7 @@ class UpdateActivityLogMessage
         $this->dateTime = $dateTime;
     }
 
-    public function getUpdatedItem()
+    public function getUpdatedItem(): object
     {
         return $this->updatedItem;
     }
@@ -40,5 +46,50 @@ class UpdateActivityLogMessage
     public function getActionName(): string
     {
         return $this->actionName;
+    }
+
+    public function isCreateLogWithoutChanges(): bool
+    {
+        return $this->createLogWithoutChanges;
+    }
+
+    public function getUser(): ?array
+    {
+        return $this->userInfo;
+    }
+
+    public function setUser(?array $userInfo): void
+    {
+        $this->userInfo = $userInfo;
+    }
+
+    public function getRequestUrl(): ?string
+    {
+        return $this->requestUrl;
+    }
+
+    public function setRequestUrl(?string $requestUrl): void
+    {
+        $this->requestUrl = $requestUrl;
+    }
+
+    public function getNormalizedItem(): array
+    {
+        return $this->normalizedItem;
+    }
+
+    public function setNormalizedItem(array $normalizedItem): void
+    {
+        $this->normalizedItem = $normalizedItem;
+    }
+
+    public function getObjectId()
+    {
+        return $this->getUpdatedItem()->getId();
+    }
+
+    public function getClassName(): string
+    {
+        return ClassUtils::getClass($this->getUpdatedItem());
     }
 }
