@@ -1,6 +1,6 @@
 <?php
 
-namespace Locastic\ActivityLog\Bridge\Elasticsearch;
+namespace Locastic\Loggastic\Bridge\Elasticsearch;
 
 use Elastic\Elasticsearch\Client;
 use Elastic\Elasticsearch\ClientBuilder;
