@@ -1,11 +1,12 @@
 <?php
 
-namespace Locastic\ActivityLog\Bridge\Elasticsearch\Index;
+namespace Locastic\Loggastic\Bridge\Elasticsearch\Index;
 
-use Locastic\ActivityLog\Bridge\Elasticsearch\Context\ElasticsearchContextInterface;
+use Locastic\Loggastic\Bridge\Elasticsearch\Context\ElasticsearchContextInterface;
 
 interface ElasticsearchIndexConfigurationInterface
 {
     public function getActivityLogIndexConfig(ElasticsearchContextInterface $elasticsearchContext): array;
+
     public function getCurrentDataTrackerIndexConfig(ElasticsearchContextInterface $elasticsearchContext): array;
 }
