@@ -1,10 +1,9 @@
 <?php
 
-namespace Locastic\ActivityLog\Bridge\Elasticsearch\Index;
+namespace Locastic\Loggastic\Bridge\Elasticsearch\Index;
 
-use Locastic\ActivityLog\Bridge\Elasticsearch\Context\ElasticsearchContextFactoryInterface;
-use Locastic\ActivityLog\Bridge\Elasticsearch\ElasticsearchClient;
-use Elastic\Elasticsearch\Exception\ClientResponseException;
+use Locastic\Loggastic\Bridge\Elasticsearch\Context\ElasticsearchContextFactoryInterface;
+use Locastic\Loggastic\Bridge\Elasticsearch\ElasticsearchClient;
 
 class ElasticsearchIndexFactory implements ElasticsearchIndexFactoryInterface
 {
@@ -21,7 +20,7 @@ class ElasticsearchIndexFactory implements ElasticsearchIndexFactoryInterface
 
     public function recreateActivityLogIndex(string $className): void
     {
-        $elasticContext = $this->elasticsearchContextFactory->createFromClassName($className);
+        $elasticContext = $this->elasticsearchContextFactory->create($className);
         $params = $this->elasticsearchIndexConfiguration->getActivityLogIndexConfig($elasticContext);
 
         $this->deleteIndex($elasticContext->getActivityLogIndex());
@@ -30,7 +29,7 @@ class ElasticsearchIndexFactory implements ElasticsearchIndexFactoryInterface
 
     public function recreateCurrentDataTrackerLogIndex(string $className): void
     {
-        $elasticContext = $this->elasticsearchContextFactory->createFromClassName($className);
+        $elasticContext = $this->elasticsearchContextFactory->create($className);
         $params = $this->elasticsearchIndexConfiguration->getCurrentDataTrackerIndexConfig($elasticContext);
 
         $this->deleteIndex($elasticContext->getCurrentDataTrackerIndex());
@@ -39,7 +38,7 @@ class ElasticsearchIndexFactory implements ElasticsearchIndexFactoryInterface
 
     public function createActivityLogIndex(string $className): void
     {
-        $elasticContext = $this->elasticsearchContextFactory->createFromClassName($className);
+        $elasticContext = $this->elasticsearchContextFactory->create($className);
 
         $params = $this->elasticsearchIndexConfiguration->getActivityLogIndexConfig($elasticContext);
 
@@ -48,7 +47,7 @@ class ElasticsearchIndexFactory implements ElasticsearchIndexFactoryInterface
 
     public function createCurrentDataTrackerLogIndex(string $className): void
     {
-        $elasticContext = $this->elasticsearchContextFactory->createFromClassName($className);
+        $elasticContext = $this->elasticsearchContextFactory->create($className);
 
         $params = $this->elasticsearchIndexConfiguration->getCurrentDataTrackerIndexConfig($elasticContext);
 
@@ -59,9 +58,9 @@ class ElasticsearchIndexFactory implements ElasticsearchIndexFactoryInterface
     {
         try {
             $this->elasticsearchClient->getClient()->indices()->delete(['index' => $index]);
-        } catch (ClientResponseException $e) {
+        } catch (\Exception $e) {
             // don't throw exception if index doesn't exist
-            if ($e->getCode() !== 404) {
+            if (404 !== $e->getCode()) {
                 throw $e;
             }
         }
