@@ -1,11 +1,11 @@
 <?php
 
-namespace Locastic\ActivityLog\DataProvider;
+namespace Locastic\Loggastic\DataProvider;
 
-use Locastic\ActivityLog\Bridge\Elasticsearch\Context\ElasticsearchContextFactoryInterface;
-use Locastic\ActivityLog\Bridge\Elasticsearch\ElasticsearchService;
-use Locastic\ActivityLog\Model\CurrentDataTracker;
-use Locastic\ActivityLog\Model\CurrentDataTrackerInterface;
+use Locastic\Loggastic\Bridge\Elasticsearch\Context\ElasticsearchContextFactoryInterface;
+use Locastic\Loggastic\Bridge\Elasticsearch\ElasticsearchService;
+use Locastic\Loggastic\Model\CurrentDataTracker;
+use Locastic\Loggastic\Model\CurrentDataTrackerInterface;
 
 class CurrentDataTrackerProvider implements CurrentDataTrackerProviderInterface
 {
@@ -20,10 +20,10 @@ class CurrentDataTrackerProvider implements CurrentDataTrackerProviderInterface
 
     public function getCurrentDataTrackerByClassAndId(string $className, $objectId): ?CurrentDataTrackerInterface
     {
-        $elasticContext = $this->elasticsearchContextFactory->createFromClassName($className);
+        $elasticContext = $this->elasticsearchContextFactory->create($className);
 
         $body = [
-            'query' => ['term' => ["objectId" => $objectId]]
+            'query' => ['term' => ['objectId' => $objectId]],
         ];
 
         //todo move class to config
