@@ -1,12 +1,16 @@
 <?php
 
-namespace Locastic\ActivityLog\Factory;
+namespace Locastic\Loggastic\Factory;
 
-use Locastic\ActivityLog\Model\ActivityLogInterface;
-use Locastic\ActivityLog\Model\CurrentDataTrackerInterface;
+use Locastic\Loggastic\Message\ActivityLogMessageInterface;
+use Locastic\Loggastic\Model\ActivityLogInterface;
+use Locastic\Loggastic\Model\CurrentDataTrackerInterface;
 
 interface ActivityLogFactoryInterface
 {
-    public function createActivityLog($id, string $resourceClass, string $action, array $data = []): ActivityLogInterface;
+    public function create(): ActivityLogInterface;
+
+    public function createFromActivityLogMessage(ActivityLogMessageInterface $activityLogMessage, ?array $data = []): ActivityLogInterface;
+
     public function createCurrentDataTracker($item, $normalizedData): CurrentDataTrackerInterface;
 }
