@@ -2,14 +2,14 @@
 
 namespace Locastic\Loggastic;
 
-use Locastic\Loggastic\DependencyInjection\LocasticActivityLogExtension;
+use Locastic\Loggastic\DependencyInjection\LocasticLoggasticExtension;
 use Symfony\Component\DependencyInjection\Extension\Extension;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
-class LocasticActivityLogsBundle extends Bundle
+class LocasticLoggasticBundle extends Bundle
 {
     public function getContainerExtension(): Extension
     {
-        return new LocasticActivityLogExtension();
+        return new LocasticLoggasticExtension();
     }
 }
