@@ -1,14 +1,11 @@
 <?php
 
-namespace Locastic\ActivityLog\Metadata\LoggableContext;
+namespace Locastic\Loggastic\Metadata\LoggableContext;
 
 class LoggableContextCollection implements \IteratorAggregate, \Countable
 {
     private array $loggableContextCollection;
 
-    /**
-     * @param string[] $loggableContextCollection
-     */
     public function __construct(array $loggableContextCollection = [])
     {
         $this->loggableContextCollection = $loggableContextCollection;
