@@ -1,14 +1,14 @@
 <?php
 
-namespace Locastic\ActivityLog\Tests\IntegrationTest\Metadata;
+namespace Locastic\Loggastic\Tests\IntegrationTest\Metadata;
 
-use Locastic\ActivityLog\Metadata\LoggableContext\Factory\LoggableContextCollectionFactoryInterface;
-use Locastic\ActivityLog\Tests\Fixtures\DummyBlogPost;
-use Locastic\ActivityLog\Tests\Fixtures\DummyCategory;
-use Locastic\ActivityLog\Tests\Fixtures\DummyPhoto;
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+use Locastic\Loggastic\Metadata\LoggableContext\Factory\LoggableContextCollectionFactoryInterface;
+use Locastic\Loggastic\Tests\Fixtures\App\Model\DummyBlogPost;
+use Locastic\Loggastic\Tests\Fixtures\App\Model\DummyCategory;
+use Locastic\Loggastic\Tests\Fixtures\App\Model\DummyPhoto;
+use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 
-class LoggableContextCollectionFactoryTest extends WebTestCase
+class LoggableContextCollectionFactoryTest extends KernelTestCase
 {
     public function testCreate()
     {
@@ -21,19 +21,19 @@ class LoggableContextCollectionFactoryTest extends WebTestCase
             'groups' => [
                 'dummy_category_log',
                 'dummy_photo_log',
-                ]
+                ],
         ], $loggableContextCollection[DummyCategory::class]);
 
         self::assertEquals([
             'groups' => [
                 'dummy_photo_log',
-            ]
+            ],
         ], $loggableContextCollection[DummyPhoto::class]);
 
         self::assertEquals([
             'groups' => [
                 'dummy_blog_post_log',
-            ]
+            ],
         ], $loggableContextCollection[DummyBlogPost::class]);
     }
 }
